--- ninchatsdk/src/main/java/com/ninchat/sdk/helper/translation/NinchatTranslation.kt
@@ -1,8 +1,0 @@
-package com.ninchat.sdk.helper.translation
-
-import com.ninchat.sdk.helper.siteconfigparser.NinchatSiteConfig
-
-class NinchatTranslation {
-    var ninchatSiteConfig: NinchatSiteConfig? = null
-    
-}
